Class Groups library
The backend is modular to, in theory, enable a constant time and a variable
time backend (not to mention to experiment with various backends). Currently, a
malachite (pure Rust) and gmp (C FFI'd by rug) backend are provided.

Compression is implemented.

The class group determination code always uses malachite. This is *fine* as
that code should only be called during setup, not during protocols proper.