Add helper for to_le_bits
The helper maps from bool to Choice and attempts to limit the risk of
side-channels. This is done on a best-effort basis with no guarantees, as all
zeroization is.

Also adds the proofs module to the public API to reduce how many items are at
the top-level. The crate isn't benefited by having multiple hash-to-primes at
the top-level.