Add a constant-time class groups implementation on the stack
Comparable to CryptoPrimesStack, this implementation panics if a request is
made which exceeds its capacity (a short-coming the heap version does not
have). This implementation has comparable performance to the heap version
*despite* fixing itself to 2688-bit coordinates (whereas the heap-version gets
away with 128-bit numbers).

It also avoids allocations during composition so it ideally shouldn't leave
copies laying in memory (unzeroized/of questionable zeroization) even without
`zalloc`, and would be a candidate for a no-std work.

The next steps are further evaluation of what algorithms are taking the most
time (GCD, reduction, etc) and the question of if more efficient algorithms
exist when targetting a constant-time implementation. The reduction algorithm
as written, terminated as soon as it reached its goal and was allowed to have a
suboptimal worst-case bound so long as it had an optimal average case. An
algorithm which does more work per step, but less steps and total work overall,
should be looked for.