Correct the amount of bits shortened to for CryptoBigintHeapElement
This implies we may be iterating more than necessary? We do have a `- 1` which
may be redundant _with_ the `div_ceil` (at least, for some configurations). It
could also be with how our log2 estimations aren't perfectly precise.

Since this should be one extra round at worst (of thousands), it isn't
currently worth ensuring the perfect efficiency in detail so long as it's
correct (which it now is).