Revert "Optimize the second CRT invocation during composition"
This reverts commit 1fa29bf08e61ed2a4bd9a2248ddb8e624d568423.

The commit had logic premised on `B_mu / e` being coprime to
`A_1 * A_2 / e**2`. The issue is that `e` is not the product of the lists
`common_factors(B_mu, A_1), common_factors(B_mu, A_2)` yet the intersection.
If `e` and `A_1` shared a factor not present in `A_2`, it wouldn't be present
in `e`, causing `e, A_1, A` all to not be coprime.