Iterate for log2 b, not for log2 b = log2 sqrt discriminant
This doubles the runtime when the prior amount of iterations had yet to fail
(even with full-size discriminants). It's accordingly unfortunate/annoying as
it doubles the runtime.

This is the proven bound however. The prior bound was because
`a <= sqrt(|discriminant|)` for a reduced form, and that same logic was applied
to `b`. This isn't the case as `b` may be arbitrarily small.

The first reduction algorithm outputs `b <= 2a`, and once `b` is less than `a`,
the form is reduced. The second algorithm will perform the final reduction
steps.