Allow crypto-bigint to use larger tables
The point of limiting the table size was because the crypto-bigint
multiplication algorithm cannot simply perform a lookup against the table, yet
must iterate over very value. This means while we only pay one addition cost
per n bits, we still copy 2**n*b bytes of memory (where b is how many bytes of
memory a single element uses). This causes a break-even point where the cost of
these memory copies become more expensive than the additions saves.

2**8 was conservatively chosen as the table size limit. With some benchmarking,
the cost of a multiplication of a 2048-bit scalar and a 2**12 table is the same
as with a 2**14 table. A 2**12 bit table is notably faster than prior table
sizes however.